+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Set User-Agent header
+
 ## [0.2.3] - 2024-06-21
 
 - Dependency updates